+++ src/core/server/utility/mongo.ts
@@ -9,7 +9,8 @@ const DEFAULT_COLLECTIONS: string[] = [
     Collections.Accounts,
     Collections.Characters,
     Collections.Options,
-    Collections.Interiors
+    Collections.Interiors,
+    Collections.Vehicles
 ];
 
 export default {